+++ src/Engines/QueryBuilderEngine.php
@@ -3,6 +3,7 @@
 namespace Yajra\Datatables\Engines;
 
 use Closure;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Query\Builder;
 use Illuminate\Support\Str;
 use Yajra\Datatables\Helper;
@@ -256,14 +257,18 @@ class QueryBuilderEngine extends BaseEngine
      */
     protected function compileGlobalSearch($query, $column, $keyword)
     {
-        $column = $this->castColumn($column);
-        $sql    = $column . ' LIKE ?';
-        if ($this->isCaseInsensitive()) {
-            $sql     = 'LOWER(' . $column . ') LIKE ?';
-            $keyword = Str::lower($keyword);
-        }
+        if ($this->isSmartSearch()) {
+            $column = $this->castColumn($column);
+            $sql    = $column . ' LIKE ?';
+            if ($this->isCaseInsensitive()) {
+                $sql     = 'LOWER(' . $column . ') LIKE ?';
+                $keyword = Str::lower($keyword);
+            }
 
-        $query->orWhereRaw($sql, [$keyword]);
+            $query->orWhereRaw($sql, [$keyword]);
+        } else { // exact match
+            $query->orWhereRaw("$column like ?", [$keyword]);
+        }
     }
 
     /**
@@ -318,15 +323,25 @@ class QueryBuilderEngine extends BaseEngine
                     );
                 }
             } else {
-                $column  = $this->castColumn($column);
-                $keyword = $this->getSearchKeyword($index);
-
-                if ($this->isCaseInsensitive()) {
-                    $this->compileColumnSearch($index, $column, $keyword, false);
-                } else {
-                    $col = strstr($column, '(') ? $this->connection->raw($column) : $column;
-                    $this->compileColumnSearch($index, $col, $keyword, true);
+                if (count(explode('.', $column)) > 1) {
+                    $eagerLoads     = $this->getEagerLoads();
+                    $parts          = explode('.', $column);
+                    $relationColumn = array_pop($parts);
+                    $relation       = implode('.', $parts);
+                    if (in_array($relation, $eagerLoads)) {
+                        $column = $this->joinEagerLoadedColumn($relation, $relationColumn);
+                    }
                 }
+
+                $column          = $this->castColumn($column);
+                $keyword         = $this->getSearchKeyword($index);
+                $caseInsensitive = $this->isCaseInsensitive();
+
+                if (! $caseInsensitive) {
+                    $column = strstr($column, '(') ? $this->connection->raw($column) : $column;
+                }
+
+                $this->compileColumnSearch($index, $column, $keyword, $caseInsensitive);
             }
 
             $this->isFilterApplied = true;
@@ -362,10 +377,12 @@ class QueryBuilderEngine extends BaseEngine
     {
         if ($this->request->isRegex($i)) {
             $this->regexColumnSearch($column, $keyword, $caseSensitive);
-        } else {
+        } elseif ($this->isSmartSearch()) {
             $sql     = $caseSensitive ? $column . ' LIKE ?' : 'LOWER(' . $column . ') LIKE ?';
             $keyword = $caseSensitive ? $keyword : Str::lower($keyword);
             $this->query->whereRaw($sql, [$keyword]);
+        } else { // exact match
+            $this->query->whereRaw("$column LIKE ?", [$keyword]);
         }
     }
 
@@ -443,20 +460,39 @@ class QueryBuilderEngine extends BaseEngine
      */
     protected function joinEagerLoadedColumn($relation, $relationColumn)
     {
-        $table   = $this->query->getRelation($relation)->getRelated()->getTable();
-        $foreign = $this->query->getRelation($relation)->getQualifiedForeignKey();
-        $other   = $this->query->getRelation($relation)->getQualifiedOtherKeyName();
-        $column  = $table . '.' . $relationColumn;
-
         $joins = [];
         foreach ((array) $this->getQueryBuilder()->joins as $key => $join) {
             $joins[] = $join->table;
         }
 
-        if (! in_array($table, $joins)) {
-            $this->getQueryBuilder()
-                 ->leftJoin($table, $foreign, '=', $other);
+        $model = $this->query->getRelation($relation);
+        if ($model instanceof BelongsToMany) {
+            $pivot   = $model->getTable();
+            $pivotPK = $model->getForeignKey();
+            $pivotFK = $model->getQualifiedParentKeyName();
+
+            if (! in_array($pivot, $joins)) {
+                $this->getQueryBuilder()->leftJoin($pivot, $pivotPK, '=', $pivotFK);
+            }
+
+            $related = $model->getRelated();
+            $table   = $related->getTable();
+            $tablePK = $related->getForeignKey();
+            $tableFK = $related->getQualifiedKeyName();
+
+            if (! in_array($table, $joins)) {
+                $this->getQueryBuilder()->leftJoin($table, $pivot . '.' . $tablePK, '=', $tableFK);
+            }
+        } else {
+            $table   = $model->getRelated()->getTable();
+            $foreign = $model->getQualifiedForeignKey();
+            $other   = $model->getQualifiedOtherKeyName();
+            if (! in_array($table, $joins)) {
+                $this->getQueryBuilder()->leftJoin($table, $foreign, '=', $other);
+            }
         }
+
+        $column = $table . '.' . $relationColumn;
 
         return $column;
     }