+++ src/config/config.php
@@ -2,6 +2,7 @@
 
 return [
     'search' => [
+        'smart'            => true,
         'case_insensitive' => true,
         'use_wildcards'    => false,
     ],